fix bugs task login not response data type json when login false
